// lib/api/model/modelLoonstrook.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Jeffries Homeapp API
 * Backend REST API for Jeffries Homeapp.
 * OpenAPI spec version: 1.0
 */
import type { ModelLoonstrookComponenten } from './modelLoonstrookComponenten';
import type { ModelLoonstrookCumulatieven } from './modelLoonstrookCumulatieven';
import type { ModelLoonstrookOrtDetail } from './modelLoonstrookOrtDetail';

export interface ModelLoonstrook {
  amt_zeerintensief?: number;
  bruto_betaling?: number;
  bruto_inhouding?: number;
  componenten?: ModelLoonstrookComponenten;
  cumulatieven?: ModelLoonstrookCumulatieven;
  eju_bedrag?: number;
  extra_uren_bedrag?: number;
  geimporteerd_op?: string;
  id?: string;
  jaar?: number;
  loonheffing?: number;
  netto?: number;
  ort_detail?: ModelLoonstrookOrtDetail;
  ort_totaal?: number;
  parttime_factor?: number;
  pensioenpremie?: number;
  periode?: number;
  periode_label?: string;
  reiskosten?: number;
  salaris_basis?: number;
  schaalnummer?: string;
  toeslag_balansvlf?: number;
  trede?: string;
  type?: string;
  user_id?: string;
  uurloon?: number;
  vakantietoeslag?: number;
}